requirements.txt
python-docx
python-dotenv
google-genai
pdfplumber
pillow
